refactor(125/WS1.4): extract auth.rs + retire test_auth_shim (#125)
First move in the src/server/mod.rs decomposition series (WS1.4).

WHAT:
- Extract src/server/auth.rs: auth_middleware, create_session handler,
  CORS-origin predicates (is_allowed_loopback_origin + helpers), policy
  predicates (is_auth_exempt_path, accepts_query_token), ct_eq,
  SessionStore, load_or_generate_api_token. All moved verbatim; mod.rs
  now references them via auth:: (pub(super) visibility).
- mod.rs: -775 lines. auth.rs: 796 lines. Net code unchanged.

THE PINNED CONDITION — authorize() pure function + shim retirement:
- New authorize(path, method, header_token, query_token, &api_token,
  &sessions, now) -> Result<(), StatusCode> captures the ENTIRE auth
  decision (OPTIONS bypass → exempt → bearer durable/session → query
  session-only → 401) as a pure, directly-callable function.
- auth_middleware is now a thin extraction layer: it pulls the header
  and query tokens into owned Strings (no request borrows survive across
  next.run(req).await) and delegates to authorize(). No decision logic.
- DELETED test_auth_shim, auth_matrix_router, matrix_request,
  matrix_ok_stub — the test-only control-flow mirror is gone. All 9+2
  auth-matrix tests now call authorize() directly, the same function
  production runs. The durable-in-query→401 property (and the entire
  matrix) is now guaranteed by production code, not a parallel copy.
- CORS tests rewritten from router-based (ACAO-header) to pure-predicate
  tests on is_allowed_loopback_origin / _str (the router shim they
  depended on is gone). cors_origin_allows_only_literal_loopback_ips,
  gui_requires_auth_but_accepts_query_token_bootstrap moved to auth.rs.
- SessionStore, ct_eq, extract_bearer, extract_query_token tests
  unchanged (already pure). 3 new extraction-helper tests added.

CHARACTERIZATION GUARD (byte-identical):
- x0x routes (text + JSON) diffed against tests/snapshots/server-routes.*
  → IDENTICAL (P0 snapshot unchanged).
- route_set_matches_registry + manifest_matches_registry green.
- All 1383 lib tests pass; no behavior change.

Gates: fmt, clippy --all-targets --all-features -- -D warnings, full lib
test suite (1383 passed), registry trio (api_coverage + api_manifest +
parity_cli), route snapshot diff — all green.

Refs #125.
